+++ lib/features/settings/widgets/help_screen.dart
@@ -805,7 +805,7 @@ class HelpScreen extends ConsumerWidget {
             context,
             Icons.code,
             'Term Translation',
-            'These prompts may need adjustment depending on the model used -- Translate the [language] term "[term]" into natural English using this sentence for context: "[sentence]". Return 1 or 2 distinct English translations only. Do not repeat the same translation. Do not include the original [language] term. If there is only one good translation, return only one. Output only the translation words or short phrases, separated by a comma.',
+            'These prompts may need adjustment depending on the model used -- Translate the [language] term "[term]" into natural English using this sentence for context: "[sentence]". Return 1 or 2 distinct English translations only. Do not repeat the same translation. Do not include the original [language] term. If there is only one good translation, return only one. Output only the translation words, separated by a comma.',
           ),
           _buildControlItem(
             context,