+++ docs/ChangeLog.md
@@ -1,5 +1,5 @@
 # Bug Fixes
--
+- Removed accidental 'or short phrases' from AI term translation
 
 # New Feature - Custom Themes!
 -